+++ CHANGELOG.md
@@ -29,5 +29,6 @@
 - Добавлено восстановление исходных Combobox values через `form.reset()`.
 - Добавлена настройка `autoHighlight` для автоматической активации первого option.
 - Добавлена настройка `autoSelect` для подтверждения active option при blur.
+- Добавлена настройка `disabledItemsFocusable` для CollectionController.
 - Добавлены unit- и DOM-тесты.
 - Добавлена проверка покрытия через `npm run coverage`.